Moved gdt_init from boot.rs to gdt.rs
