+++ src/kernel/Cargo.toml
@@ -14,6 +14,7 @@ path = "src/bin.rs"
 
 [dependencies]
 kernel_sync = { path = "../sync/" }
+# spin = "0.9.8"
 
 [dependencies.lazy_static]
 version = "1.5.0"
