+++ src/kernel/src/arch/x86/mod.rs
@@ -1,8 +1,7 @@
-pub mod boot;
-pub mod exceptions;
 pub mod gdt;
 pub mod idt;
-pub mod paging;
+
+/* -------------------------------------- */
 
 // TODO: Look at file structure & add docs
 #[doc(hidden)]
@@ -10,6 +9,12 @@ pub mod cpu;
 #[doc(hidden)]
 pub mod diagnostics;
 #[doc(hidden)]
+pub mod exceptions;
+#[doc(hidden)]
 pub mod io;
 #[doc(hidden)]
+pub mod paging;
+#[doc(hidden)]
 pub mod target;
+
+/* -------------------------------------- */
